refactor: domian structure for modules y dinamis initialization con associtations
