+++ core/controllers/users.js
@@ -1,6 +1,10 @@
 import { ResponseHelper as response } from "../helpers/response.js";
-import dotenv from 'dotenv';
-import { validateUserIdParameters, validateUserParameters, validateUserUpdateParameters } from "../validations/userValidations.js";
+import dotenv from "dotenv";
+import {
+  validateUserIdParameters,
+  validateUserParameters,
+  validateUserUpdateParameters,
+} from "../validations/userValidations.js";
 import * as userService from "../services/userService.js";
 import { ValidationError } from "yup";
 
@@ -11,6 +15,7 @@ export const getUsers = async () => {
     const users = await userService.getUsers();
     return response.success(users);
   } catch (error) {
+    console.log("aca esta el error", error);
     return response.badRequest(error);
   }
 };
@@ -45,7 +50,7 @@ export const createUser = async (event) => {
     const newUser = await userService.createUser(userData);
     return response.success({
       message: "User created successfully",
-      user: newUser
+      user: newUser,
     });
   } catch (error) {
     if (error instanceof ValidationError) {
@@ -56,7 +61,7 @@ export const createUser = async (event) => {
     }
     return response.badRequest("Error creating user: " + error.message);
   }
-}
+};
 
 export const updateUser = async (event) => {
   try {
@@ -67,8 +72,10 @@ export const updateUser = async (event) => {
 
     const updatedUser = await userService.updateUser(id, userData);
 
-    return response.success({ message: `Usuario con id ${id} actualizado`, user: updatedUser });
-
+    return response.success({
+      message: `Usuario con id ${id} actualizado`,
+      user: updatedUser,
+    });
   } catch (error) {
     if (error instanceof ValidationError) {
       return response.badRequest({
@@ -86,7 +93,6 @@ export const deleteUser = async (event) => {
     await validateUserIdParameters({ id });
     await userService.deleteUser(id);
     return response.success({ message: `Usuario con id ${id} eliminado` });
-
   } catch (error) {
     if (error instanceof ValidationError) {
       return response.badRequest({
@@ -107,7 +113,7 @@ export const activateUser = async (event) => {
 
     return response.success({
       message: `Usuario con id ${id} activado`,
-      user
+      user,
     });
   } catch (error) {
     if (error instanceof ValidationError) {
@@ -129,7 +135,7 @@ export const deactivateUser = async (event) => {
 
     return response.success({
       message: `Usuario con id ${id} desactivado`,
-      user
+      user,
     });
   } catch (error) {
     if (error instanceof ValidationError) {
