+++ core/models/access/index.js
@@ -1,0 +1,12 @@
+import { RoleModel, RoleFields } from "./RoleModel.js";
+import { PermissionModel, PermissionFields } from "./PermissionModel.js";
+import { AccessLogModel, AccessLogFields } from "./AccessLogModel.js";
+
+export {
+  RoleModel,
+  RoleFields,
+  PermissionModel,
+  PermissionFields,
+  AccessLogModel,
+  AccessLogFields,
+};
