+++ core/models/application/index.js
@@ -1,0 +1,12 @@
+import { ApplicationModel, ApplicationFields } from "./ApplicationModel.js";
+import { DynamicRouteModel, DynamicRouteFields } from "./DynamicRouteModel.js";
+import { MenuItemModel, MenuItemFields } from "./MenuItemModel.js";
+
+export {
+  ApplicationModel,
+  ApplicationFields,
+  DynamicRouteModel,
+  DynamicRouteFields,
+  MenuItemModel,
+  MenuItemFields,
+};
