+++ core/models/modules/index.js
@@ -1,0 +1,15 @@
+import { CustomModuleModel, CustomModuleFields } from "./CustomModuleModel.js";
+import { ModuleModel, ModuleFields } from "./ModuleModel.js";
+import {
+  OrganizationModuleModel,
+  OrganizationModuleFields,
+} from "./OrganizationModuleModel.js";
+
+export {
+  CustomModuleModel,
+  CustomModuleFields,
+  ModuleModel,
+  ModuleFields,
+  OrganizationModuleModel,
+  OrganizationModuleFields,
+};
