+++ core/models/organization/index.js
@@ -1,0 +1,3 @@
+import { OrganizationModel, OrganizationFields } from "./OrganizationModel.js";
+
+export { OrganizationModel, OrganizationFields };
