+++ core/models/time/index.js
@@ -1,0 +1,3 @@
+import { TimeTrackerModel, TimeTrackerFields } from "./TimeTrackerModels.js";
+
+export { TimeTrackerModel, TimeTrackerFields };
