+++ core/models/user/index.js
@@ -1,0 +1,15 @@
+import { UserModel, UserFields } from "./UserModel.js";
+import { UserInfoModel, UserInfoFields } from "./UserInfoModel.js";
+import { UserWorkInfoModel, UserWorkInfoFields } from "./UserWorkInfoModel.js";
+import { UserRoleModel, UserRoleFields } from "./UserRoleModel.js";
+
+export {
+  UserModel,
+  UserFields,
+  UserInfoModel,
+  UserInfoFields,
+  UserWorkInfoModel,
+  UserWorkInfoFields,
+  UserRoleModel,
+  UserRoleFields,
+};
