+++ core/package.json
@@ -1,26 +1,26 @@
 {
-  "name": "core",
-  "version": "1.0.0",
-  "description": "",
-  "type": "module",
-  "scripts": {
-    "test": "jest",
-    "deploy_dev": "serverless deploy --stage dev",
-    "deploy_function_dev": "serverless deploy function --stage dev --function",
-    "dev": "serverless offline --stage dev --noAuth --noTimeout --reloadHandler --verbose"
-  },
-  "dependencies": {
-    "@aws-sdk/client-cognito-identity-provider": "^3.749.0",
-    "aws-sdk": "^2.1692.0",
-    "bcryptjs": "^3.0.0",
-    "dotenv": "^16.4.7",
-    "express": "^4.19.2",
-    "mysql2": "^3.12.0",
-    "sequelize": "^6.37.5",
-    "serverless-http": "^3.2.0",
-    "yup": "^1.6.1"
-  },
-  "devDependencies": {
-    "serverless-offline": "^14.4.0"
-  }
+	"name": "core",
+	"version": "1.0.0",
+	"description": "",
+	"type": "module",
+	"scripts": {
+		"test": "jest",
+		"deploy_dev": "serverless deploy --stage dev",
+		"deploy_function_dev": "serverless deploy function --stage dev --function",
+		"dev": "serverless offline --stage dev --noAuth --noTimeout --reloadHandler --verbose"
+	},
+	"dependencies": {
+		"@aws-sdk/client-cognito-identity-provider": "^3.749.0",
+		"aws-sdk": "^2.1692.0",
+		"bcryptjs": "^3.0.0",
+		"dotenv": "^16.4.7",
+		"express": "^4.19.2",
+		"mysql2": "^3.12.0",
+		"sequelize": "^6.37.5",
+		"serverless-http": "^3.2.0",
+		"yup": "^1.6.1"
+	},
+	"devDependencies": {
+		"serverless-offline": "^14.4.0"
+	}
 }
